:hammer: Songs play when album clicked
Added js to make the according song play when an album is clicked.
The song stops if the same album is clicked again.
If another album is clicked then the previous song stops and the new plays.
The title of the description also changes with the according title of the song and the artist.
The zodiac circle rotates to show the according zodiac sign.
Background gradient and some minor CSS fixes.